refactor: rename to event-bus-websockets-server
