+++ test/common/event-bus/websockets/event-bus-websockets.spec.js
@@ -3,7 +3,7 @@
 import 'module-alias/register';
 import { assert, expect } from 'chai';
 import { spy, stub } from 'sinon';
-import { SocketEventBus } from '@event-bus/websockets/event-bus-websockets';
+import { SocketEventBus } from '@event-bus/websockets/event-bus-websockets-server';
 
 describe('SocketEventBus', function() {
     let eventBus = new SocketEventBus();
